* Big change: Add kernel-global "KBDev" keycode enum. * I8042Prt: Use the kbdev keycodes instead of own scan codes. * VideoTerminal: Watch for kbdev keycodes instead of ps/2 scan codes.
(Note: I couldn't figure out a better name yet, sorry)

+++ user/VideoTerminal/source/terminal.h
@@ -26,6 +26,6 @@ BSTATUS CreateIOLoopThreads();
 BSTATUS WaitOnIOLoopThreads();
 
 // Inter-module operation
-void TranslateKeyCode(char* Input, uint8_t KeyCode);
+void TranslateKeyCode(char* Input, uint16_t KeyCode);
 void TerminalWrite(const char* Buffer, size_t Length);
 void TerminalGetDimensions(int* Width, int* Height);
